[AXON-715][Rovo Dev] Fire 'Deep plan shown' telemetry event (#613)
